+++ Changelog.md
@@ -5,6 +5,7 @@
 - Use `&'a [T]` instead of `&'a Vec<T>` in `DataOrSuffStat`
 - Use `&[T]` instead of `&Vec<T>` in `GewekeTester`
 - Renamed `BesselIvError` variants
+- Implement `Rv<usize>` for `Poisson`
 
 ## 0.12.0
 - Added Inverse X^2 distribution